Split converter tests to separate module per converter and drop tests/data/*tdb
